move to local and global components
