--- packages/components/src/index.ts
@@ -1,4 +1,0 @@
-export * from './Header/index.js'
-export * from './Toast/index.js'
-export * from './shared/index.js'
-export * from './Buttons/index.js'
