+++ packages/components/src/local/Form/index.ts
@@ -1,0 +1,1 @@
+export { HdForm, type HdFormProps } from './HdForm.js'
